gateway/fault_inject/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-can>=4.0
paho-mqtt==2.1.0
docker>=7.0
